Proposal to add `oci-artifact` as pURL type
The current pURL specification has a proposed `Docker` entry that can be
used when referencing docker container images. However, Docker Hub is
not the only registry where container images might be stored.
Additionally, a container image is not the only type of artifact that
may be stored in an oci registry.

The SPDX community is designing a means to uniquely identify artifacts
that an SBoM is associated within. In SPDX 3.0 specifically, this will
be the `Artifact URL` property which refers to a unique artifact. In
order to represent all cloud native artifacts, there needs to be an
entry in the pURL spec that can represent more than just docker
container images. Adding the `oci-artifact` entry would specifically
enable:

- a means to uniquely identify any artifact, regardless of what registry
it may currently be in. (Use cases: importing an artifact from a public
registry to a private registry, and promotion within and across private
registries)
- reference to the same global identifier when describing
identical oci artifacts in different SBoMs (Use cases: two different
companies describing the same oci artifact using the same global
identifier in two different SPDX documents)

Signed-off-by: Rose Judge <rjudge@vmware.com>